compass working, not working with turning left and right however. Rotations only use encoder board

+++ SDPArduinoCode/compass.h
@@ -10,6 +10,6 @@
 
 	void setupCompass();
 	float updateCompass();
-	void printCompass(int headingDegrees);
+	void printCompass(float headingDegrees);
 
 #endif
